chore: refactor server to grpc_app
