+++ src/bentoml_cli/worker/grpc_dev_api_server.py
@@ -52,6 +52,13 @@ import click
     default=None,
     help="CA certificates file",
 )
+@click.option(
+    "--protocol-version",
+    type=click.Choice(["v1", "v1alpha1"]),
+    help="Determine the version of generated gRPC stubs to use.",
+    default="v1",
+    show_default=True,
+)
 def main(
     bento_identifier: str,
     host: str,
@@ -64,6 +71,7 @@ def main(
     ssl_certfile: str | None,
     ssl_keyfile: str | None,
     ssl_ca_certs: str | None,
+    protocol_version: str,
 ):
     import psutil
 
@@ -96,12 +104,13 @@ def main(
 
         asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())  # type: ignore
 
-    from bentoml._internal.server import grpc
+    from bentoml._internal.server import grpc_app as grpc
 
     grpc_options: dict[str, t.Any] = {
         "bind_address": f"{host}:{port}",
         "enable_reflection": enable_reflection,
         "enable_channelz": enable_channelz,
+        "protocol_version": protocol_version,
     }
     if max_concurrent_streams is not None:
         grpc_options["max_concurrent_streams"] = int(max_concurrent_streams)
